% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itsa.postest.R
\name{itsa.postest}
\alias{itsa.postest}
\title{Post-estimating ITSA models}
\usage{
itsa.postest(model = NULL, no.plots = FALSE, alpha = 0.05)
}
\arguments{
\item{model}{define ITSA model results object, default is output from itsa.model function}

\item{no.plots}{logical, specify whether function should present post-estimation plots (default is FALSE)}

\item{alpha}{define alpha level for test statistics, default is 0.05}
}
\value{
if no.plots is not forced, will return a table to the console and two plots to the plot tab
}
\description{
Function to produce post-estimation plots from the itsa.model function.
}
\details{
This function runs and reports post-estimation tests on fits from the itsa.model function, and generates four plots.

Main tests are whether two key ANCOVA assumptions are met, and an additional autocorrelation test for the time series framework.

The Shaprio-Wilks test examines the residuals from the fitted model for normality. A p-value less than alpha indicates abnormal residuals.

The Levene's Test makes sure that there are equal variances between the treated groups. A p-value less than alpha indicates hetrogenous variances.

A QQ-Norm and Boxplot are generated with the test results overlaid (respectively), with a Residual v Fitted and Autocorrelation Function Plot also generated.

Default is to generate plots and summary table, but plots may be overriden using no.plots argument.

Default alpha value for post-estimation statistics is 0.05, test results will suggest potential presence of problems at higher values (and also at higher levels relative to a user-inputted alpha), but user discretion is needed (examined in tandam with the Residuals v Fitted plot).

See 'itsa.model' documentation for further information.
}
\keyword{analysis}
\keyword{interrupted}
\keyword{of}
\keyword{post-estimation}
\keyword{series,}
\keyword{time}
\keyword{variance,}
